+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-* Now the actionbar is also offset along with other elements.
-* As much duplicate code as possible has been removed.
+* Now the actionbar is also offset along with other elements